CommandType is "long" type not "short" type
See https://api.libreoffice.org/docs/idl/ref/namespacecom_1_1sun_1_1star_1_1sdb_1_1CommandType.html
+ discussion here:
http://document-foundation-mail-archive.969070.n3.nabble.com/Type-of-CommandType-short-or-long-tt4299416.html

Important remark from Stephan:
"...the data type corresponding to UNOIDL LONG (i.e., 32 bit) is xs:int, not xs:long"

+++ sw/source/uibase/dbui/mmconfigitem.cxx
@@ -517,7 +517,7 @@ void  SwMailMergeConfigItem_Impl::ImplCommit()
             case 23: pValues[nProp] <<= m_sMailPassword;        break;
             case 24 :pValues[nProp] <<= m_aDBData.sDataSource;           break;
             case 25 :pValues[nProp] <<= m_aDBData.sCommand;              break;
-            case 26 :pValues[nProp] <<= static_cast<short>(m_aDBData.nCommandType);   break;
+            case 26 :pValues[nProp] <<= m_aDBData.nCommandType;   break;
             case 27 :pValues[nProp] <<= m_sFilter; break;
             case 28 :pValues[nProp] <<= m_aSavedDocuments; break;
             case 29: pValues[nProp] <<= m_bIsEMailSupported; break;
